Django-2-ORM done

+++ Django-2-ORM/d42/ex04/views.py
@@ -15,7 +15,8 @@ def init(request):
         query = sql.SQL("""
             CREATE TABLE IF NOT EXISTS ex04_movies (
                 title VARCHAR(64) UNIQUE NOT NULL,
-                episode_nb INTEGER PRIMARY KEY CHECK (episode_nb > 0),
+                episode_nb INTEGER PRIMARY KEY,
+                opening_crawl TEXT,
                 director VARCHAR(32) NOT NULL,
                 producer VARCHAR(128) NOT NULL,
                 release_date DATE NOT NULL
@@ -31,76 +32,67 @@ def init(request):
 def populate(request):
     
     movies = [
-        (1, 'The Phantom Menace', 'George Lucas', 'Rick McCallum', '1999-05-19'),
-        (2, 'Attack of the Clones', 'George Lucas', 'Rick McCallum', '2002-05-16'),
-        (3, 'Revenge of the Sith', 'George Lucas', 'Rick McCallum', '2005-05-19'),
-        (4, 'A New Hope' , 'George Lucas, Gary Kurtz', 'Rick McCallum', '1977-05-25'),
-        (5, 'The Empire Strikes Back', 'Irvin Kershner', 'Gary Kurtz, Rick McCallum', '1980-05-17'),
-        (6, 'Return of the Jedi', 'Richard Marquand', 'Howard G. Kazanjian, George Lucas, Rick McCallum', '1983-05-25'),
-        (7, 'The Force Awakens', 'J. J. Abrams, Kathleen Kennedy', 'J. J. Abrams, Bryan Burk', '2015-12-11'),
+        (1, 'The Phantom Menace', None, 'George Lucas', 'Rick McCallum', '1999-05-19'),
+        (2, 'Attack of the Clones', None, 'George Lucas', 'Rick McCallum', '2002-05-16'),
+        (3, 'Revenge of the Sith', None, 'George Lucas', 'Rick McCallum', '2005-05-19'),
+        (4, 'A New Hope', None, 'George Lucas', 'Gary Kurtz, Rick McCallum', '1977-05-25'),
+        (5, 'The Empire Strikes Back', None, 'Irvin Kershner', 'Gary Kurtz, Rick McCallum', '1980-05-17'),
+        (6, 'Return of the Jedi', None, 'Richard Marquand', 'Howard G. Kazanjian, George Lucas, Rick McCallum', '1983-05-25'),
+        (7, 'The Force Awakens', None, 'J. J. Abrams', 'Kathleen Kennedy, J. J. Abrams, Bryan Burk', '2015-12-11'),
     ]
     result_msg = []
     
-    try:
-        query = """
-            INSERT INTO ex04_movies (episode_nb, title, director, producer, release_date)
-            VALUES (%s, %s, %s, %s, %s);
-        """
-        with connection.cursor() as cursor:
-            
-            for movie in movies:
-                episode_nb = movie[0]
-                
-                cursor.execute("SELECT episode_nb FROM ex04_movies WHERE episode_nb = %s;", [episode_nb])
-                already_exist = cursor.fetchone()
-                
-                if not already_exist:
-                    cursor.execute(query, movie)
-                    result_msg.append("OK")
-    except Exception as e:
-        return HttpResponse(f"Error: {e}")
-    if result_msg:
-        return HttpResponse("<br>".join(result_msg))
-    else:
-        return HttpResponse("No data available")
+    query = """
+        INSERT INTO ex04_movies (episode_nb, title, opening_crawl, director, producer, release_date)
+        VALUES (%s, %s, %s, %s, %s, %s);
+    """
+    for movie in movies:
+        try:
+            with connection.cursor() as cursor:
+                cursor.execute(query, movie)
+            result_msg.append("OK")
+        except Exception as e:
+            result_msg.append(f"Error: {e}")
+    return HttpResponse("<br>".join(result_msg))
 
 def display(request):
     
     try:
         query = sql.SQL("""
-            SELECT episode_nb, title, director, producer, release_date
+            SELECT episode_nb, title, opening_crawl, director, producer, release_date
             FROM ex04_movies;
         """)
         with connection.cursor() as cursor:
             cursor.execute(query)
             results = cursor.fetchall()
+        if not results:
+            return HttpResponse('No data available')
         return render(request, 'ex04/display.html', {'results': results})
         
     except Exception as e:
         return HttpResponse(f'No data available')
 
 def remove(request):
-    choices = []
     try:
         with connection.cursor() as cursor:
-            cursor.execute("SELECT title from ex04_movies ORDER BY title;")
+            cursor.execute("SELECT title FROM ex04_movies ORDER BY title;")
             choices = [row[0] for row in cursor.fetchall()]
-    except psycopg2.Error as e:
-        print(f"Database Error: {e}")
-    
+    except psycopg2.Error:
+        return HttpResponse("No data available")
+
+    if not choices:
+        return HttpResponse("No data available")
+
     if request.method == 'POST':
-        form = DropDown(request.POST)
+        form = DropDown(choices, request.POST)
         if form.is_valid():
             title_to_remove = form.cleaned_data['films_titles']
             try:
                 with connection.cursor() as cursor:
                     cursor.execute("DELETE FROM ex04_movies WHERE title = %s;", [title_to_remove])
-            except psycopg2.Error as e:
+            except psycopg2.Error:
                 return HttpResponse("No data available")
         return HttpResponseRedirect('/ex04/remove')
-    else:
-        form = DropDown()
-    return render(request, 'ex04/remove.html', {
-        'form': form,
-        'choices': choices,
-    })
+
+    form = DropDown(choices)
+    return render(request, 'ex04/remove.html', {'form': form})
